feat: add the home-cards component
